requirements.txt
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
bleach==6.0.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
dnspython==2.4.2
docker-pycreds==0.4.0
einops==0.6.1
email-validator==2.0.0.post2
executing==1.2.0
fastapi==0.103.1
filelock==3.12.4
fonttools==4.42.1
fsspec==2023.9.1
gitdb==4.0.10
GitPython==3.1.36
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
huggingface-hub==0.17.1
idna==3.4
ipykernel==6.25.2
ipython==8.15.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
loguru==0.7.2
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numpy==1.25.2
orjson==3.9.7
packaging==23.1
pandas==2.1.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==2.3.0
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
setproctitle==1.3.2
six==1.16.0
sklearn==0.0.post9
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
text-unidecode==1.3
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
transformers==4.33.2
typing_extensions==4.7.1
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
wandb==0.15.10
watchfiles==0.20.0
wcwidth==0.2.6
webencodings==0.5.1
websockets==11.0.3